新增分类接口

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -1,19 +1,23 @@
 package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
+import com.sky.mapper.SetmealMapper;
 import com.sky.service.CategoryService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.beans.beancontext.BeanContext;
 import java.time.LocalDateTime;
+import java.util.List;
 
 /**
  * @projectName: sky-take-away
@@ -29,6 +33,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private DishMapper dishMapper;
+
+    @Autowired
+    private SetmealMapper setmealMapper;
 
     @Override
     public void save(CategoryDTO categoryDTO) {
@@ -50,6 +60,16 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategoryById(long id) {
+        // 判断当前分类是否关联到菜品
+        Integer count = dishMapper.countByCategoryId(id);
+        if(count > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        // 判断当前分类是否关联到套餐
+        count = setmealMapper.countByCategoryId(id);
+        if(count > 0){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
         categoryMapper.delete(id);
     }
 
@@ -60,5 +80,20 @@ public class CategoryServiceImpl implements CategoryService {
                 .id(id)
                 .build();
         categoryMapper.update(category);
+    }
+
+    @Override
+    public void update(CategoryDTO categoryDTO) {
+        Category category = Category.builder()
+                .updateTime(LocalDateTime.now())
+                .updateUser(BaseContext.getCurrentId())
+                .build();
+        BeanUtils.copyProperties(categoryDTO,category);
+        categoryMapper.update(category);
+    }
+
+    @Override
+    public List<Category> list(Integer type) {
+        return categoryMapper.list(type);
     }
 }
